requirements.txt
pillow
tensorflow
streamlit
numpy